Importing .xml settings into build.gradle
I'm personally offended by the maker of this plugin for not allowing to load .xml cfg files and not working version-mechanisms.
Oh and some settings are missing.

I'll probably ditch the plugin and just put a launch4j downloading script in the repo (like the windows-jre and probably also the linux-jre in the future)