Added launch4j downloading script
Figured I'd ditch the plugin because it sucks quite badly for some reason.
A more elegant solutions would be to create a fork and fix it. I could also do that for InnoSetup and the platform-specific JRE's.

Even better would be to dynamically download the JRE upon first launch, however that would require some native windows/mac/linux coding.
Could be a nice idea for a project though...